+++ src/config.rs
@@ -1,11 +1,19 @@
-use std::{env, fmt};
-
+#[cfg(not(feature = "blocking"))]
+use governor::{
+    clock::{QuantaClock, QuantaInstant},
+    middleware::NoOpMiddleware,
+    state::{InMemoryState, NotKeyed},
+    Quota, RateLimiter,
+};
 #[cfg(feature = "blocking")]
 use reqwest::blocking::{Client, RequestBuilder, Response};
 use reqwest::header::USER_AGENT;
 #[cfg(not(feature = "blocking"))]
 use reqwest::{Client, RequestBuilder, Response};
 use reqwest::{Method, Url};
+use std::{env, fmt};
+#[cfg(not(feature = "blocking"))]
+use std::{num::NonZeroU32, sync::Arc, time::Duration};
 
 use crate::{error::types::ErrorResponse, Error, Result};
 
@@ -14,6 +22,8 @@ pub struct Config {
     pub(crate) api_key: String,
     pub(crate) base_url: Url,
     pub(crate) client: Client,
+    #[cfg(not(feature = "blocking"))]
+    limiter: Arc<RateLimiter<NotKeyed, InMemoryState, QuantaClock, NoOpMiddleware<QuantaInstant>>>,
 }
 
 impl Config {
@@ -28,11 +38,31 @@ impl Config {
 
         let env_user_agent = format!("{}/{}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"));
 
+        // ==== Rate limiting is a non-blocking thing only ====
+        #[cfg(not(feature = "blocking"))]
+        let rate_limit_per_sec = env::var("RATE_LIMIT")
+            .unwrap_or_else(|_| "10".to_owned())
+            .parse::<u32>()
+            .expect("env variable `RATE_LIMIT` should be a valid u32");
+
+        #[cfg(not(feature = "blocking"))]
+        let quota = Quota::with_period(Duration::from_millis(1100))
+            .expect("Valid quota")
+            .allow_burst(
+                NonZeroU32::new(rate_limit_per_sec).expect("Rate limit is a valid non zero u32"),
+            );
+
+        #[cfg(not(feature = "blocking"))]
+        let limiter = Arc::new(RateLimiter::direct(quota));
+        // ====================================================
+
         Self {
             user_agent: env_user_agent,
             api_key: api_key.to_owned(),
             base_url: env_base_url,
             client,
+            #[cfg(not(feature = "blocking"))]
+            limiter,
         }
     }
 
@@ -51,7 +81,15 @@ impl Config {
 
     #[maybe_async::maybe_async]
     pub async fn send(&self, request: RequestBuilder) -> Result<Response> {
+        #[cfg(not(feature = "blocking"))]
+        {
+            let jitter =
+                governor::Jitter::new(Duration::from_millis(10), Duration::from_millis(50));
+            self.limiter.until_ready_with_jitter(jitter).await;
+        }
+
         let request = request.build()?;
+
         let response = self.client.execute(request).await?;
 
         match response.status() {