+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 ### Added
 
 - `DomainChanges.tls`
+- Rate limiting (default 10/s), configurable with `RATE_LIMIT` env variable (async only)
 
 ### Changed